Move C and therefore CGO out of shared package
The following moves any C and therefore CGO code out of the shared
package. This then allows other projects (like Juju) to still use the
client package whilst themselves also use CGO to compile. Prior to this
change, we end up with either missing headers or redefinition errors
when attempting to use the client.

If there is a better place to put the files, then let me know.
